requirements.txt
PySide6>=6.9.0
torch~=2.8.0
torchvision~=0.23.0
timm>=1.0.19,<2.0
einops>=0.6.0
numpy>=1.26,<3
safetensors==0.4.2
Pillow>=9.4.0